Move code from header files to cpp files
