--- src/Exceptions/InvalidCommand/InvalidCommandException.cpp
@@ -1,1 +1,0 @@
-#include "InvalidCommandException.hpp"
