--- src/Exceptions/NotFound/NotFoundException.cpp
@@ -1,2 +1,0 @@
-#include "NotFoundException.hpp"
-
