--- src/Exceptions/Socket/SocketException.cpp
@@ -1,1 +1,0 @@
-#include "SocketException.hpp"
