--- src/Exceptions/WrongFormat/WrongFormatException.cpp
@@ -1,1 +1,0 @@
-#include "WrongFormatException.hpp"
