--- src/Format/Ascii/AsciiFormat.cpp
@@ -1,1 +1,0 @@
-#include "AsciiFormat.hpp"
