--- src/Format/Flatbuffers/FlatbuffersFormat.cpp
@@ -1,1 +1,0 @@
-#include "FlatbuffersFormat.hpp"
