--- src/Format/MGMT/MGMTFormat.cpp
@@ -1,1 +1,0 @@
-#include "MGMTFormat.hpp"
