+++ src/Transport/SocketContainer/SocketContainer.hpp
@@ -11,25 +11,11 @@
 class SocketContainer {
 
 public:
-  SocketContainer& register_socket(std::shared_ptr<AbstractSocket> socket) {
-    Packet::Type packet_type = socket->format()->packet_type();
-    m_sockets.emplace(packet_type, std::move(socket));
+  // Register given socket in container
+  SocketContainer& register_socket(std::shared_ptr<AbstractSocket> socket);
 
-    return *this;
-  }
-
-  bool send(std::unique_ptr<Packet::AbstractPacket> packet) {
-    Packet::Type type = packet->current_type();
-
-    auto it = m_sockets.find(type);
-    if (it == m_sockets.end()) {
-      throw Exceptions::NotFoundException("Can't find socket in SocketContainer for given packet type.");
-    }
-    std::shared_ptr<AbstractSocket> socket = it->second;
-
-    std::vector<uint8_t> data = packet->serialize();
-    return socket->send(data);
-  };
+  // Send a packet using the matching registered socket
+  bool send(std::unique_ptr<Packet::AbstractPacket> packet);
 
 private:
   std::unordered_map<Packet::Type, std::shared_ptr<AbstractSocket>> m_sockets;
